Change quantity_per_pack to use whole numbers instead of decimals
- Update migration to use unsignedInteger instead of decimal
- Update model cast to integer
- Update controller validation to require integer with min:1
- Update view inputs to use step=1 and default value of 1
- Remove decimal support from all quantity inputs

+++ database/migrations/2025_11_18_023045_create_packing_material_blueprints_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
                 ->constrained('items')
                 ->onDelete('restrict')
                 ->name('pmb_material_item_fk');
-            $table->decimal('quantity_per_pack', 10, 3);
+            $table->unsignedInteger('quantity_per_pack');
             $table->timestamps();
 
             $table->unique(
